minor changes + start fixing tree
